requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
scipy>=1.11
pandas>=2.1
pyarrow>=14.0
opencv-python>=4.9
albumentations>=1.4
pyyaml>=6.0
tqdm>=4.66
requests>=2.31
streamlit>=1.33
matplotlib>=3.8
pillow>=10.2
pytest>=8.0
